bullseye: armbian-resize-filesystem fails on first run due to missing fdisk (#1992)
* armbian-resize-filesystem fails on first run due to missing fdisk

The `fdisk` package is missing in the `minimal` image, in bullseye at leat. Maybe other, future minimal images as well.
This will cause the `armbian-resize-filesystem` script to fail on first boot, which contains several
`fdisk` references. This could be fixed in the `bullseye` specific section, but since fdisk is so
essential for Armbian, it's probably best to list it in the general, mandatory packages to make sure it's available in every image in any case.