Add session chooser to lightdm menu (#2341)
Right now, you simply cannot choose between different DE's installed on the system. The session indicator uses very little space, and doesn't clutter the greeter.
I noticed it was missing when trying to integrate a Kodi session. It will be useful in the future also, when we incorporate more DE's than Xfce