RK3328 GPU DMC OPP tables (#3088)
* [ RK3328 ] boost GPU voltages to avoid starving Memory controller

The RK3328 has the dynamic memory controller and GPU on the same power rail.
Both are able to be dynamically clocked, both have opp tables.

 - DMC min voltage:  1.075V
 - GPU min voltage:  0.950V

On boards with only the GPU enabled, they memory controller can be under-voltaged.
On boards attempting to use both, the disagreeing opps result in the GPU failing to init

This patch applies a mV range opp to the GPU and fixed opps to the DMC on rk3328-roc-cc
to allow everything to get alongin all possible devfreq situations.

* [ rockchip64 ] clean up disabled patches

* [ rock64 ] add mali supply and remove unused 5V supply

- New Mali opps will keep from starving ram controller

- extra supply was causing dmesg errors.

As it shares enable with another supply and both were marked "always on"
I removed the second.  Neither are referenced by any ports/devices.