JetHome: JetHub D1 fix patch: sdio: update JetHub D1 regulators for SDIO (#3389)
Fix the dts to match board's reference design:
- update vddio_boot regulator to 3.3v (Wi-Fi SDIO module)
- add vccq_1v8 regulator with 1.8v for eMMC SDIO

In the first revision of JetHub D1 the vccq_1v8 regulator was 3.3v.
All installed eMMC modules were tested to work in HS200 mode at 3.3v
supply voltage. In the next revisions of the device eMMC will be
powered with 1.8v according to the standard.