Fix armbian header (#4144)
* bsp: quote pattern passed to grep

This fixes a regression from #4013

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>

* bsp: refactor DISTRIBUTION_CODENAME logic

There's no need to repeat the check for /etc/armbian-distribution-status
over and over again, let's just use an if-statement instead.

Similarly, let's not repeat the logic to look up DISTRIBUTION_STATUS in
/etc/armbian-distribution-status; we can just do it at the end instead.

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>

* bsp: only fallback to /etc/os-release if /etc/lsb-release failed

This fixes a regression from #3619

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>

* bsp: only fallback to "lsb_release -c" if all else failed

Debian Bookwork and Sid both neither have the debian-edu-config package
by default (which provodes /etc/lsb-release in Debian), nor does their
version of base-files (12.2) contain the VERSION_CODENAME field.

So we need another fallback. We *do* have lsb_release, which can provide
this information as well. So let's do that, but add a check as well in
case neither the lsb-release nor the lsb-release-minimal packages is
installed on a user's system.

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>