prevent armbian-install from running on aml-s9xx-box (#6247)
* aml-s9xx-box uses the u-boot shipped with the TV Box instead of installing
a new u-boot.  Because of this, armbian-install will make the box unbootable
if it is run to install to emmc.  This change prevents running
armbian-install on aml-s9xx-box boards

Changes to be committed:
	modified:   packages/bsp/common/usr/sbin/armbian-install

* prevent armbian-install from running on aml-s9xx-box take 2

Per suggestion of @viraniac remove armbian-install from the board bsp
via the board config file

Changes to be committed:
	modified:   config/boards/aml-s9xx-box.tvb
	modified:   packages/bsp/common/usr/sbin/armbian-install