actions: Move concurrency canceling to 2nd job for PR build
On every PR, a workflow is started to check if artifacts should be
built. This happens not only once, but many times, e.g. for every
selected reviewer. Since the workflow has `cancel-in-progress`
enabled, workflows are started and immediately cancelled by the
next one, resulting in many notifications.
Move the `cancel-in-progress` concurrency policy to the second job
which starts only after a check is done if the 'Build' label is even
active on the PR. This should greatly reduce "Workflow cancelled"
notifications via GitHub and email (if enabled by the user).

Also make wording more clear.