aml-s9xx-box: Remove u-boot and update generated extlinux.conf
 Changes to be committed:
	modified:   config/boards/aml-s9xx-box.tvb
	modified:   packages/bsp/aml-s9xx-box/boot/extlinux/extlinux.conf.template

The aml-s9xx-box builds use the native TV Box u-boot.  I've had as a
todo item to figure out how to not build and include an unused u-boot.
In a recent PR for a different board the opposite was being done
(adding a u-boot where a board didn't previously ship one), thus
I then had the info to remove it here.
This also changes the generated extlinux.conf to include a fdt line
by adding the boot_fdt_file to the board config file.  This reduces
the effort needed to manually modify the extlinux.conf.