firstlogin: bring up wifi device before scanning
Scanning accesspoints requires an activated wifi device.
Networkd does not activate it (only NetworkManager does so).
As a result, firstlogin is not able to configure and use wifi on minimal
images, as only Server and Desktop images have NetworkManeger installed.
Activating the wifi device using "ip link set ${WIFI_DEVICE} up" before
scanning fixes this.
(And can safely be done - there is no problem activating it multiple
times, as it is the case with NetworkManager installed.)