sunxi-6.12: BUGFIX: Remove the incomplete patch.
When starting a GPU with a power domain, it is necessary
to additionally control the switching on and off of the clock
in the GPU driver.
The driver is currently waiting for a response when it is
turned on, but it is not there, and this causes the OS to freeze.

The complete set will be added later after testing.