extensions: Add TI's repo and install specified packages in k3-*
Texas Instruments maintains a custom apt repository [0] that contains:
    * tools like k3conf, which run on K3 devices
    * TI's versions of upstream packages (such as mesa)
    * out-of-tree drivers and firmware for graphics, wifi etc

Therefore, add TI's custom repository as the highest priority repository
in the filesystem. Doing this ensures that if apt finds a version of a
package that exists in both upstream Debian and the TI repository, it
picks the latter.

Additionally, introduce K3_PACKAGES variable to store a list of packages
that should be installed by-default in a K3 image. Initialize it to hold
TI's CC33xx packages.

Also set EXTRAWIFI to "no" in `current` image.

[0] https://github.com/TexasInstruments/ti-debpkgs

Co-authored-by: Suhaas Joshi <s-joshi@ti.com>
Signed-off-by: Suhaas Joshi <s-joshi@ti.com>