So I'm not completely convinced about CONFIG_ARM64_ERRATUM_858921, but when doing my initial changes [enabled a bunch of PHYs, made NFS/SMB etc modules... random stuff], I somehow enabled some errata fixes, and this one showed up in dmesg.
The real star of the show I'm pretty sure is CONFIG_PCS_MTK_USXGMII.
Note, it cannot be a module, but on this platform we probably wouldn't
want it to be anyway. When I say "it can't", I'm saying it breaks the
build during the linker phase for vmlinux. Probably worth fixing, but
I'm unclear where to start, other than it's probably a Kconfig
dependency problem.

CONFIG_I2C_SLAVE_EEPROM came along for the ride b/c it possibly made
sense with SFPs. But I haven't tried a test build without it yet.

For the record: this works with both my AQS-107-B0C2-CX [an 802.3bz
SFP+] and my FS.com SFP-10GSR-85 "generic" SFP+.